Rename 'components/MetroBarChart.vue' to 'components/index/CardsReference/Metro/Chart.vue'
